test-python-MironBerch/solution/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
dj-database-url==2.1.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.31
PyJWT==2.8.0
